refactor: move task aliases scenarios to TaskAliases
